put the tests where Django will actually run them!
